requirements.txt
python-telegram-bot==20.8
docker==6.1.3
psutil==5.9.5
python-dotenv==1.0.0
urllib3==2.2.1 